requirements.txt
opencv-python
torch
torchvision
transformers
Pillow
requests
ultralytics
fastapi
uvicorn
python-multipart